Return form of DOLISTS is omitted.
In the first place, the RETURN form of DOLIST macro is in vain.  Since
DOLIST cannot introduce any variable except the iterator, we can never
return any useful value unless it is wrapped by a form that introduce
some variable like LET. And if a DOLIST form is wrapped by a LET form
for example, then we can return the value from the LET form instead.